+++ src/debug.c
@@ -26,6 +26,7 @@ int main(int argc, char **argv) {
         {desktop, "desktop"},
         {gtk_theme, "gtk_theme"},
         {icon_theme, "icon_theme"},
+        {cursor_theme, "cursor_theme"},
         {shell, "shell"},
         {login_shell, "login_shell"},
         {term, "term"},