+++ src/info/info.h
@@ -19,6 +19,8 @@ int gtk_theme(char *dest);
 
 int icon_theme(char *dest);
 
+int cursor_theme(char *dest);
+
 int shell(char *dest);
 
 int login_shell(char *dest);